+++ src/router.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { createBrowserRouter, Navigate } from "react-router-dom";
-import SignupForm from "./components/signupForm";
+import SignupForm from "./pages/signupForm";
 
 const router = createBrowserRouter([
   {
